+++ services/backend/toolwarehouse-service/src/main/java/com/veteroch4k/toolwarehouse/models/ToolReservedEvent.java
@@ -1,5 +1,5 @@
 package com.veteroch4k.toolwarehouse.models;
 
-public record ToolReservedEvent(int orderId) {
+public record ToolReservedEvent(Long orderId) {
 
 }
