+++ services/backend/toolwarehouse-service/src/main/java/com/veteroch4k/toolwarehouse/models/commands/RequiredTools.java
@@ -1,6 +1,6 @@
 package com.veteroch4k.toolwarehouse.models.commands;
 
-public record RequiredTools(int toolType,
-                            int quantity) {
+public record RequiredTools(Long toolType,
+                            Long quantity) {
 
 }
