+++ services/backend/toolwarehouse-service/src/main/java/com/veteroch4k/toolwarehouse/models/commands/ToolReservationCommand.java
@@ -2,6 +2,6 @@ package com.veteroch4k.toolwarehouse.models.commands;
 
 import java.util.List;
 
-public record ToolReservationCommand(int orderId, List<RequiredTools> tools, int factoryId) {
+public record ToolReservationCommand(Long orderId, List<RequiredTools> tools, Long factoryId) {
 
 }
