+++ services/backend/toolwarehouse-service/src/main/java/com/veteroch4k/toolwarehouse/repositories/FactoryToolsRepository.java
@@ -4,8 +4,8 @@ import com.veteroch4k.toolwarehouse.models.FactoryTools;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FactoryToolsRepository extends JpaRepository<FactoryTools, Integer> {
+public interface FactoryToolsRepository extends JpaRepository<FactoryTools, Long> {
 
-  List<FactoryTools> findAllByFactoryId(int factoryId);
+  List<FactoryTools> findAllByFactoryId(Long factoryId);
 
 }
