+++ services/backend/toolwarehouse-service/src/main/java/com/veteroch4k/toolwarehouse/repositories/ToolTypeRepository.java
@@ -4,6 +4,6 @@ import com.veteroch4k.toolwarehouse.models.ToolType;
 import jakarta.persistence.criteria.CriteriaBuilder.In;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ToolTypeRepository extends JpaRepository<ToolType, Integer> {
+public interface ToolTypeRepository extends JpaRepository<ToolType, Long> {
 
 }
